// src/model/interface/usagePoint/TrafoV3DTO.ts
export interface TrafoV3DTO {
  id: number;
  customer_kind_1_count: number;
  customer_kind_2_count: number;
  customer_kind_3_count: number;
  customer_kind_4_count: number;
  customer_kind_5_count: number;
  customer_kind_6_count: number;
  lat: number;
  lon: number;
  type_id: number;
  usage_point_count: number;
}